requirements.txt
confluent-kafka==2.5.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pandas==2.2.2
streamlit==1.36.0
SQLAlchemy==2.0.32
